Added custome methods for device_go,
when TDP limit is enabled, the device will change mode to custom (purple), and set the TDP using the wmi interface. This is only done when legacy mode is disabled in config.

When TDP limit is disabled, the device will change mode to balanced (white led).

Currently requires the acpi_call kernel module to be loaded.
Using sudo modprobe acpi_call